Update: name refactor and eslint
