+++ .eslintrc.js
@@ -10,12 +10,16 @@ module.exports = {
   extends: [
     // https://github.com/vuejs/eslint-plugin-vue#priority-a-essential-error-prevention
     // consider switching to `plugin:vue/strongly-recommended` or `plugin:vue/recommended` for stricter rules.
-    'plugin:vue/essential'
+    'plugin:vue/recommended',
+    'plugin:prettier/recommended',
+    'eslint:recommended'
   ],
   // required to lint *.vue files
   plugins: [
     'vue'
   ],
   // add your custom rules here
-  rules: {}
+  rules: {
+    "prettier/prettier": false
+  }
 }
