+++ layouts/default.vue
