+++ package.json
@@ -19,8 +19,10 @@
   "devDependencies": {
     "babel-eslint": "^8.2.1",
     "eslint": "^4.15.0",
+    "eslint-config-prettier": "^3.0.1",
     "eslint-friendly-formatter": "^3.0.0",
     "eslint-loader": "^1.7.1",
+    "eslint-plugin-prettier": "^2.6.2",
     "eslint-plugin-vue": "^4.0.0",
     "node-sass": "^4.9.3",
     "sass-loader": "^7.1.0"
